refactor(core-p2p): don't test private methods in the peer guard
