--- packages/core-p2p/src/court/index.ts
@@ -1,1 +1,0 @@
-export * from "./guard";
