+++ packages/core-p2p/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./monitor";
 export * from "./peer";
-export * from "./court";
 export * from "./plugin";
 export * from "./network-state";
 export * from "./utils";
