+++ packages/core-p2p/src/interfaces.ts
@@ -2,3 +2,11 @@ export interface IAcceptNewPeerOptions {
     seed?: boolean;
     lessVerbose?: boolean;
 }
+
+export interface IOffence {
+    number: number;
+    period: string;
+    reason: string;
+    weight: number;
+    critical?: boolean;
+}
