+++ packages/core-p2p/src/peer.ts
@@ -5,8 +5,8 @@ import AJV from "ajv";
 import socketCluster from "socketcluster-client";
 import util from "util";
 import { config as localConfig } from "./config";
-import { guard } from "./court";
 import { PeerPingTimeoutError, PeerStatusResponseError, PeerVerificationFailedError } from "./errors";
+import { guard } from "./guard";
 import { PeerVerificationResult, PeerVerifier } from "./peer-verifier";
 import { replySchemas } from "./reply-schemas";
 import { SocketErrors } from "./socket-server/constants";
